fix(clippy): resolve pre-existing clippy warnings for CI -D warnings
- config.rs: replace vec![] with array slices (useless_vec)
- ollama.rs: replace constant assert with let binding (assertions_on_constants)
- main.rs: use struct literal for Config (field_reassign_with_default)
- context_smoke.rs: replace format!("{big}") with big.clone() (useless_format)

+++ crates/zoid-provider/src/ollama.rs
@@ -591,7 +591,7 @@ mod tests {
     #[test]
     fn default_local_num_ctx_clears_zoid_fixed_overhead() {
         // DEFAULT_LOCAL_NUM_CTX is a compile-time constant >= 32768.
-        let _: usize = DEFAULT_LOCAL_NUM_CTX;
+        let _: u32 = DEFAULT_LOCAL_NUM_CTX;
     }
 
     fn ctx_req() -> CompletionRequest {
